feat: add docs to all operations

+++ routes/recommender_routes.py
@@ -1,6 +1,6 @@
 import os
 from flask import Blueprint, request, jsonify
-from operations.recommender import check_recommendation_objects, get_recommendations, restrictiveness, accessibility, catalog_coverage, visibility, controversy, global_controversy
+from operations.recommender import check_recommendation_objects, get_recommendations, restrictiveness, excluding_restrictiveness, accessibility_table, accessibility, catalog_coverage, visibility, controversy, global_controversy
 
 recommender_bp = Blueprint('recommender_bp', __name__,
                            url_prefix='/api/v1/recommendations')
@@ -113,9 +113,42 @@ def restrictiveness_route():
     else:
         return 'No file uploaded'
 
+@recommender_bp.route('/analysis/excluding-restrictiveness', methods=['POST'])
+def exclude_restrictiveness_route():
 
-@recommender_bp.route('/analysis/accessibility', methods=['POST'])
-def accessibility_route():
+    # Get files
+    model = request.files['model']
+    products = request.files['products']
+    features = request.form.getlist('features')
+
+    # split by comma
+    features = [x.strip() for x in features[0].split(',')]
+
+    # Check if files are provided
+    if model.filename != '' and products.filename != '' and len(features) != 0:
+
+        # Save files
+        model.save(os.path.join(MODEL_FOLDER, model.filename))
+        products.save(os.path.join(PRODUCT_FOLDER, products.filename))
+
+        # Result
+        result = excluding_restrictiveness(os.path.join(MODEL_FOLDER, model.filename), os.path.join(
+            PRODUCT_FOLDER, products.filename), features)
+
+        # Remove files
+        os.remove(os.path.join(MODEL_FOLDER, model.filename))
+        os.remove(os.path.join(PRODUCT_FOLDER, products.filename))
+
+        # Return result
+        return str(result)
+
+    # If no file is provided
+    else:
+        return 'No file uploaded'
+
+
+@recommender_bp.route('/analysis/accessibility-table', methods=['POST'])
+def accessibility_table_route():
 
     # Get files
     model = request.files['model']
@@ -129,8 +162,38 @@ def accessibility_route():
         products.save(os.path.join(PRODUCT_FOLDER, products.filename))
 
         # Result
-        results, _ = accessibility(os.path.join(MODEL_FOLDER, model.filename), os.path.join(
+        results, _ = accessibility_table(os.path.join(MODEL_FOLDER, model.filename), os.path.join(
             PRODUCT_FOLDER, products.filename))
+
+        # Remove files
+        os.remove(os.path.join(MODEL_FOLDER, model.filename))
+        os.remove(os.path.join(PRODUCT_FOLDER, products.filename))
+
+        # Return result
+        return str(results)
+
+    # If no file is provided
+    else:
+        return 'No file uploaded'
+
+@recommender_bp.route('/analysis/accessibility', methods=['POST'])
+def accessibility_route():
+
+    # Get files
+    model = request.files['model']
+    products = request.files['products']
+    product = request.form.getlist('product')
+
+    # Check if files are provided
+    if model.filename != '' and products.filename != '':
+
+        # Save files
+        model.save(os.path.join(MODEL_FOLDER, model.filename))
+        products.save(os.path.join(PRODUCT_FOLDER, products.filename))
+
+        # Result
+        results = accessibility(os.path.join(MODEL_FOLDER, model.filename), os.path.join(
+            PRODUCT_FOLDER, products.filename), product[0])
 
         # Remove files
         os.remove(os.path.join(MODEL_FOLDER, model.filename))
